using size_t instead of off_t
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@3193 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/C/omeis/File.c
@@ -279,7 +279,7 @@ int FinishFile (FileRep *myFile) {
 
 
 /*
-  UploadFile (char *filename, off_t size)
+  UploadFile (char *filename, size_t size)
   Makes new rep file in 'Files' of the specified size.
   copies filename parameter to OID.info
   Reads stdin, writing to the file.
